refactor(bhwi): isolate BitBox common adapters
- keep BitBox commands, responses, and transmits device-native
- move shared command and response mappings into the common layer
- preserve core and async behavior with focused adapter coverage
