+++ bhwi/src/common/adapters.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "bitbox")]
+mod bitbox;
 mod coldcard;
 mod jade;
 mod ledger;
